refactor: move shared DTO structs to /pkg/contracts
