+++ internal/web/apiclient/service.go
@@ -2,6 +2,8 @@ package api
 
 import (
 	"context"
+
+	"github.com/dan-nicholls/danlovesto.run/pkg/contracts"
 )
 
 type StatsService struct {
@@ -12,80 +14,28 @@ func NewStatsService(client *Client) *StatsService {
 	return &StatsService{ c: client }
 }
 
-type Info struct {
-	Uptime string `json:"uptime"`
-	Version string `json:"version"`
-}
-
-type Summary struct {
-	TotalRuns int `json:"total_runs"`
-	TotalDistance float64 `json:"total_distance"`
-	TotalHours int `json:"total_hours"`
-	TotalClimbed int `json:"total_climbed"`
-}
-
-type YearStats struct {
-	TotalDistance float64 `json:"totalDistance"`
-	AvgDistance float64 `json:"avgDistance"`
-}
-
-type Day struct {
-	Date string `json:"date"`
-	Distance float64 `json:"distance"`
-	Level int `json:"level"`
-}
-
-type Year struct {
-	Year int `json:"year"`
-	Days []Day `json:"days"`
-	Stats YearStats `json:"stats"`
-}
-
-type BucketDetails struct {
-	Scale 	string 	 `json:"scale"`  // INFO - Only supporting linear currently
-	Domain 	[2]float64 	 `json:"domain"` // [min, max]
-	Levels 	int 	 `json:"levels"`
-	Stops 	[]float64	 `json:"stops"` // The normalised values per bucket
-	Edges 	[]float64 	 `json:"edges"` // Scaled bucket values max*Stops[]
-	Labels 	[]string `json:"labels"`
-}
-
-type HeatmapData struct {
-	Today string `json:"today"`
-	Years []Year `json:"years"`
-	Buckets BucketDetails `json:"buckets"`
-}
-
-type HeatmapParams struct {
-	FromYear int // First year to include 
-	ToYear int // Last year to include
-	Unit string // Default final units (Default: km)
-	Scale string // Scale of buckets (Supported: Linear)
-	Levels int // Total number of buckets	
-}
-
-func (s *StatsService) Info(ctx context.Context) (Info, error) {
-	var i Info
+func (s *StatsService) Info(ctx context.Context) (contracts.Info, error) {
+	var i contracts.Info
 	if err := s.c.Get("/info", &i); err != nil {
-		return Info{}, err
+		return contracts.Info{}, err
 	}
 	return i, nil
 }
 
-func (s *StatsService) Summary(ctx context.Context) (Summary, error) {
-	var sum Summary
+func (s *StatsService) Summary(ctx context.Context) (contracts.Summary, error) {
+	var sum contracts.Summary
 	err := s.c.Get("/stats/summary", &sum)
 	if err != nil {
-		return Summary{}, err
+		return contracts.Summary{}, err
 	}
 	return sum, nil
 }
 
-func (s *StatsService) DailyLogs(ctx context.Context) (HeatmapData, error) {
-	var hd HeatmapData
+func (s *StatsService) DailyLogs(ctx context.Context) (contracts.HeatmapData, error) {
+	var hd contracts.HeatmapData
 	err := s.c.Get("/stats/heatmap", &hd)
 	if err != nil {
-		return HeatmapData{}, err
+		return contracts.HeatmapData{}, err
 	}
 	return hd, nil 
 }
