+++ pkg/contracts/pb.go
@@ -1,4 +1,4 @@
-package model
+package contracts 
 
 import (
 	"time"
